arch/libc: Integrate vfork into fork, and vfork directly call up_fork
