+++ libs/libc/unistd/Make.defs
@@ -53,9 +53,6 @@ endif
 
 ifeq ($(CONFIG_ARCH_HAVE_FORK),y)
 CSRCS += lib_fork.c
-ifeq ($(CONFIG_SCHED_WAITPID),y)
-CSRCS += lib_vfork.c
-endif
 endif
 
 # Add the unistd directory to the build
