+++ libs/libc/unistd/lib_fork.c
@@ -30,6 +30,9 @@
 
 #include <unistd.h>
 #include <stdio.h>
+#include <sys/wait.h>
+#include <errno.h>
+#include <debug.h>
 
 #if defined(CONFIG_ARCH_HAVE_FORK)
 
@@ -169,5 +172,67 @@ pid_t fork(void)
 
   return pid;
 }
+
+#if defined(CONFIG_SCHED_WAITPID)
+
+/****************************************************************************
+ * Public Functions
+ ****************************************************************************/
+
+/****************************************************************************
+ * Name: vfork
+ *
+ * Description:
+ *   The vfork() function is implemented based on fork() function, on
+ *   vfork(), the parent task need to wait until the child task is performing
+ *   exec or running finished.
+ *
+ * Returned Value:
+ *   Upon successful completion, vfork() returns 0 to the child process and
+ *   returns the process ID of the child process to the parent process.
+ *   Otherwise, -1 is returned to the parent, no child process is created,
+ *   and errno is set to indicate the error.
+ *
+ ****************************************************************************/
+
+pid_t vfork(void)
+{
+  int status = 0;
+  int ret;
+  pid_t pid;
+
+#ifdef CONFIG_PTHREAD_ATFORK
+  atfork_prepare();
+#endif
+  pid = up_fork();
+
+#ifdef CONFIG_PTHREAD_ATFORK
+  if (pid == 0)
+    {
+      atfork_child();
+    }
+  else
+    {
+      atfork_parent();
+    }
+#endif
+
+  if (pid != 0)
+    {
+      /* we are in parent task, and we need to wait the child task
+       * until running finished or performing exec
+       */
+
+      ret = waitpid(pid, &status, WNOWAIT);
+      if (ret < 0)
+        {
+          serr("ERROR: waitpid failed: %d\n", get_errno());
+        }
+    }
+
+  return pid;
+}
+
+#endif /* CONFIG_SCHED_WAITPID */
 
 #endif /* CONFIG_ARCH_HAVE_FORK */
